requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
requests>=2.31.0
edgartools>=5.13.1